+++ src/app/layout.tsx
@@ -2,7 +2,6 @@ import type { Metadata } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import './globals.css'
 import AlchemyProvider from '@/components/AlchemyProvider'
-import { Provider } from '@/components/ui/provider'
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -25,13 +24,11 @@ export default function RootLayout({
   children: React.ReactNode
 }>) {
   return (
-    <html lang='en' suppressHydrationWarning>
+    <html lang='en'>
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        <Provider>
-          <AlchemyProvider>{children}</AlchemyProvider>
-        </Provider>
+        <AlchemyProvider>{children}</AlchemyProvider>
       </body>
     </html>
   )